Some minor bugs fixed + explanations
For this and the previous commit, most of the code has been rewritten. I had made everything with JFormDesigner. More on how to install it and use is on the WhatsApp group.

If you want to add custom code to any element in any window, YOU HAVE TO DO IT THROUGH JFormDesigner!!! Otherwise it might get erased if you just write it in the java file.

Sorting has been done according to ALL parameters available right now.

We have to improve the Result Window to show only 1-2 options with more spec details about them.

WE HAVE TO ADD WAY MORE PHONES AS WELL!