Result panel modified but not finished.
The buttons "Next" and "Previous" doesn't do anything for now. Is pretty simple to make them to work.

We have to update the images for each phone.

Make the buttons "Next" and "Previous" to work as intended. Next is going to show the next recommanded phone. Also, we would have to order the phones in a descendent manner (higher price -> lower price). And have a max of 5 recommandations (in case we have more than 5 phones in our list).

Beside this, we just have to test again if the program filters everything good as Ayham had that issue.